вернул поле netto в отчет дубликатов

+++ src/main/kotlin/com/akopyan/exelparser/domain/database/ReportingPeriodsRepo.kt
@@ -3,6 +3,6 @@ package com.akopyan.exelparser.domain.database
 import org.springframework.data.repository.CrudRepository
 
 
-interface ReportingPeriodsRepo : CrudRepository<ReportingPeriods, Int> {
+interface ReportingPeriodsRepo : CrudRepository<ReportingPeriods, Long> {
     fun findByReportingPeriod(reportingPeriod: String): List<ReportingPeriods>
 }
